refactor: moved request logic from controller to service
